Recreation des classes garage et voiture dans des fichiers distincts
